Separate GTK test classes from object under test
The tests for GTK objects used multiple inheritance to combine
TestCase class and the class being tested. This made the code more
confusing, but also caused the GTK objects to be initialized when
loading (not executing) the test, which can cause a segfault if
X is not available.

I rewrote the tests to use composition instead of inheritance,
creating an object (self.window, etc.) in the setUp method.