qrexec-policy-agent: prepare for GuiVM
This changes qrexec-policy-agent to use a (socket-based) Qubes RPC
endpoint instead of D-Bus. Thanks to that, qrexec-policy can now
expect the guivm property and launch the confirmation prompt in
the right GuiVM.

(If GuiVM is dom0, we connect to the socket locally).

Requires QubesOS/qubes-core-admin#323 so that qubesd gives us the
guivm property.