Move all closing code to process_io.c; don't exit early
Paraphrasing Marek Marczykowski-Górecki:

1. Previously, getting exit code message immediately exited the loop in
process_child_io(), possibly leaving some data in stdin_buf. Fix this by
not exiting until stdin is also closed.

2. Closing FDs code checks for child_process_pid to avoid shutdown()
call on a socket that is not created in the
qrexec-agent/qrexec-fork-server process, but inherited from a parent
(as stdin/stdout) in qrexec-client-vm. But that's wrong because
child_process_pid is 0 in the case of a local process started by
qrexec-client-vm (not qrexec-agent/qrexec-fork-server). Fix this by
removing this condition and rely on comparing stdin/stdout to their
standard numbers (1, 0).