policy-agent: use autostart program, not systemd service
A systemd service (even in user instance) will not run as part of
graphical session, so the policy agent will not be automatically
stopped on logout / started on next login.

Instead, use an autostart file in the common package, but exit
immediately if we're not in a GuiVM.

Use a separate shell script (qrexec-policy-agent-autorun) for
performance reasons: Python application startup takes some time,
especially when it's importing GTK etc.