Fix not closed file descriptors in qubes-rpc-multiplexer
The comment says explicitly to be very careful about it, but the actual
code was wrong. When an individual command have stdin/out redirected,
shell saves (clone of) original open FD, to be restored when command
finishes. This means both of the subshells kept stdin/out open on
another FD, even though individal commands had it redirected.
This caused EOF from a service not being delivered to the client.

Fix this by connecting stdin/out to /dev/null for the whole subshells,
not just individual commands.