Handle POLLERR for stdin as well as POLLHUP
The stdin descriptor might be a pipe, which means we will get
POLLERR, not POLLHUP, when the other end closes it. If we do not
handle it, poll() will keep returning immediately until we close
the descriptor due to some other reason (e.g. receiving an EOF
from vchan, or receiving data from vchan that we fail to write).