daemon, agent: unify main loop
Use a pselect() in the main loop, unblocking the signals during
waiting, same as in process_io().

Still handle incoming vchan messages when outgoing vchan is blocked
(libvchan_buffer_space() too low). This was the case for agent, but
not for daemon.

Create a pselect_vchan() function that handles vchan quirks, but is
closer in semantics to regular pselect().