Remove false error message on vchan disconnect
The client code complains when we run with local process, and the
service closes vchan before that process exits. This can be
observed sometimes when running dnf in a VM, when it notifies dom0
about updates:

    process_io.c:181:process_io: vchan connection closed early
    (fds: -1 -1 -1, status: -1 126)

However, this behaviour is completely legitimate and in fact the
code handles it (there is a waitpid() after the loop). The code is
messy and demands more attention, but for now, silence this case
so that there is no false positive.