policy: interpret invalid requested target as @default
If the requested target does not exist, interpret it as it wasn't
provided at all. This prevents the source domain learning if a given
target exists when the policy is set to default `ask` action
(`... @anyvm ask`). In such a case, previous behavior differed based on
target existence: if it did exist, the user get a prompt, otherwise the
request is immediately refused.

Note it is still possible to write a policy that allows the source to
confirm/deny existence of arbitrary domain (for example default `allow`
action, or denying only `@default` target). But such policy would need
to be specifically configured, it does no longer apply to default
(innocently-looking) policy.

Fixes QubesOS/qubes-issues#5955