agent: do not crash on spurious MSG_SERVICE_CONNECT from the daemon
If (buggy) qrexec-daemon sends MSG_SERVICE_CONNECT for a given
connection, do not crash on it, log an error and ignore the message.
Blindly processing FD named in the message, makes qrexec-agent crash
with "pselect: Bad file descriptor" error.

Implement the fix naively - observe if the write() call fails with
EBADF. If it doesn't fail, assume the FD is the right one. This still
doesn't check if we are really waiting for the connection to establish
(it may be already accepted connection that is still running), but at
least avoid the service crash. The proper fix is needed on the daemon
side, to not send spurious messages.

QubesOS/qubes-issues#6120