daemon: do not send MSG_SERVICE_* messages twice for a given connection
Record if the service call request was already answered and send the
response only when it wasn't. The information about the request is
stored until qrexec-policy-exec (or a process talking to
qrexec-policy-daemon) terminates, so the qrexec-client needs to connect
in that time frame. This is currently guaranteed property
(qrexec-policy related process exits only after qrexec-client exits).

Refuse to send a response to the same request twice. But also refuse to
send a response for an unknown connection (not requested, or when
qrexec-policy related process is already terminated).

This is important, because qrexec-agent rely on getting exactly one
response to a given request.

With this change, if the qrexec-client is called twice for a given
request, the other side of the connection may still be started (for the
second time). In that case, it will timeout waiting for the data vchan
connection. To avoid this properly, qrexec-client call should be made by
the qrexec-daemon (perhaps even as a simple function call instead of
separate process), instead of qrexec-policy related process. But that is
a more elaborate change because of how Disposable VMs are handled.

Fixes QubesOS/qubes-issues#6120