Adjust vchan_{send,recv} error checking
libvchan_{send,recv} functions return -1 on error and data size on
success. If nonblocking more is enabled, it may also return 0.
Nonblocking more is not enabled currently, so checking return value
for -1 is correct. But if nonblocking mode is ever enabled, just
checking for -1 may have bad consequences, as it may misinterpret
failures as success.
Change the error checking to the safer option - always compare against
expected size.