daemon: fix checking qrexec-policy-daemon response
Fix handling truncated (including to 0) response from the policy daemon.
Ensure the response is NUL-terminated string even when truncated.
Otherwise, a stale 'response' buffer content (not overriden by recv())
could be confused as the actual response.

Note this wouldn't allow policy bypass anyway, because this response is
used for error reporting only. The actual connection setup is performed
by the qrexec-policy-daemon - so even if qrexec-daemon would
misinterpret policy-daemon response as "allow", the connection wouldn't
be opened.