Add a policy.EvalSimple qrexec service
The service is socket-based and is part of the policy daemon, so it
should be fairly efficient.  It is called “Simple” because it only
supports treating “ask” as “allow”.  The core evaluation code requires
that a domain_id and process_ident be provided, for which this code
reuses the same values (“dummy_id” and “0”, respectively) that the GUI
daemon provides when running in dom0.

If the service is invoked as something other than “policy.EvalSimple”,
it logs a warning message and fails.  It also fails if there is no
argument provided, or if the argument is the empty string.

The default policy for “policy.EvalSimple” includes admin-policy-ro.
GUI domains will need to be allowed to perform qubes.ClipboardPaste
checks for domains they provide GUI services to.  However, this will
need to be a separate service, as qrexec policy is not sufficiently
expressive for this.