Handle partial reads from StreamReader.read
During tests, we got the following confusing error:

```
May 23 00:58:38 dom0 qrexec-policy-daemon[1252]: b'policy.EvalSimple': invalid data from qube
```

This appears to indicate that the GUI daemon was sending bad data.
However, the GUI daemon was actually fine.  The real problem was that we
assume that `await reader.read(65536)` will not return until all 65536
bytes have been read or EOF is reached.  This is a bad assumption: it
will return when *any* bytes have been read from the socket.

In practice, we usually wind up receiving two buffers: one containing
the qrexec metadata and NUL separator, and the other containing the data
from the remote qube.  Because of the bug, we only read the metadata
buffer, and then errored out because we thought the remote qube didn’t
send any data.

The fix is to continue reading from the socket until either all 65536
bytes have been read or EOF is received.  To minimize the changes
needed, I chose to retry manually in Python, rather than using the
built-in `readuntil` method.  Using the built-in routines would be
faster and simpler, though, so it should be preferred.

I also added much better logging of invalid messages.  This should make
future debugging far simpler.