Avoid having qrexec-policy-daemon handle connections
This makes qrexec-policy-daemon stateless and ensures that restarting it
does not e.g. cause data sent over qrexec to be corrupted.  There should
also be a performance improvement.

Most of the decrease in lines of code is because lots of the Python
tests simply do not make sense any more, as they test code that no
longer exists.  Unit-testing C code is significantly more difficult and
so the current plan is to rely on integration tests.

Since it no longer spawns qrexec-client, qrexec-policy-daemon no longer
needs the remote domain ID or the request ID.  Not passing them
simplifies both the Python and C sides.  Furthermore, since the daemon
now rejects requests with either the remote domain ID or the request ID
present, old versions of qrexec-daemon will fall back to
qrexec-policy-exec.  This ensures that qrexec continues working during
upgrades.

Old versions of qrexec-daemon will still need qrexec-policy-exec to run
qrexec-client.  Add a unit test for this.

Fixes: QubesOS/qubes-issues#6628