fuzz: daemon->handle_message_from_agent()
This is the primary function parsing messages from VM, in dom0. Fuzz
this too. It requires a little change to the mockup, as the expected
handling of invalid message is exit(). Mock that with setjmp/longjmp and
jump back to the fuzzing entry point (LLVMFuzzerTestOneInput).

When handle_message_from_agent() handles incoming service
call, it normally spawns a new process that sends the request to the
policy engine. Since that part is just sending already validated
data to a python code, end fuzzing at this point. This is done by
mocking fork() and pretending it spawned a new process when actually it
did not. It does mean the daemon will quickly run out of pending call
slots (hardcoded limit of 256), but it doesn't matter at that point.

Also, rename main() function (via #define) to not conflict with real
fuzzer main().