Ensure that QSB-089 can never happen again
It turns out that gcc is able to validate that the QSB-089 fixes worked
and optimize accordingly.  __attribute__((error)) can be used to
generate a compile-time error if a call to a function remains after
optimization, and one can use this to validate that the fix to QSB-089
worked.  Ensures that if the code is vulnerable to QSB-089, it will not
compile.