Ensure that all external variables are type-checked
This requires that all objects with external linkage are declared in a
header file that is included by both the definition site and use site.
clang's -Werror=missing-variable-declarations can help catch this.  In
most cases, the problem was just a missing 'static', but the fuzzer
needs access to functions and variables that would normally be static.
Fix this by moving the declarations to fuzz.h and including that file in
daemon/qrexec-daemon.c if FUZZING_BUILD_MODE_UNSAFE_FOR_PRODUCTION is
defined.  Otherwise, the objects are declared 'static'.