Ensure consistent treatment of "QUBESRPC" followed by non-space
parse_qubes_rpc_command() would not treat this as a service call,
whereas exec_qubes_rpc_if_requested() would.  Therefore, the command
would be executed as a service call, but the usual check for
socket-based services would be skipped.  Furthermore,
exec_qubes_rpc_if_requested() would silently ignore everything after
"QUBESRPC" until the first space or the end of the string.

To avoid this inconsistent behavior, ensure that both
execute_qubes_rpc_if_requested() and parse_qubes_rpc_command() detect
this situation and fail the service call.  There are no tests for
QUBESRPC followed by a non-space so the current behavior is almost
certainly not intentional.  There are no tests for the new behavior,
either, but an error condition is very unlikely to be accidentally
depended on.