Report correct statuses for service execution failure
"Service not found" must be reported with status 127, and "service
cannot be executed" must be reported with status 125.  However, the code
had lots of bugs and would often report wrong status codes in these
error conditions.

This fixes all of the problems I could find, and adds tests to ensure
that many cases stay fixed.  This breaks the libqrexec ABI, but
thankfully the ABI that was broken is not in any release yet.