qrexec-daemon: partially validate messages from client
This is not strictly necessary, as the client is trusted, but it's still
good practice.  The validation is cheap and debugging a buggy client
without it would not be.

The validation is limited to that needed to prevent memory unsafety in
qrexec-daemon.  It's not designed to protect the agent, which can do as
much or as little validation as it likes.