Avoid allocating a big buffer for each loop iteration
There's no reason to allocate a big buffer for each loop iteration, only
to free it immediately afterwards.  Just allocate the buffer once and
reuse it.

This breaks the libqrexec ABI, but the changed functions are not used
outside of libqrexec itself.  To ensure that any problems are caught
early, move the functions to a separate header and mark them as having
hidden visibility.