Test VM => dom0 calls with skip-service-descriptor=true
The test currently fails because the command is parsed, the
configuration is loaded, and then a GUI session is waited for if needed.
Then the parsed command is destroyed, and a _new_ parsed command is
created and used for execution, _without_ loading the configuration.

This will be fixed by not destroying and re-parsing the configuration.

The negative diffstat is because the tests had a massive amount of
duplicated code, some of which is now replaced by a common function.