Avoid qrexec-client for VM -> dom0 calls
It isn't needed in this case either.  This saves an execve() and makes
it easier to change the behavior of qrexec-client without breaking
existing qrexec-daemon processes.

This also ensures that the service configuration loaded for VM -> dom0
service calls affects the 'struct qrexec_parsed_command' used for
service execution, which fixes a bug: skip-service-descriptor=true
was ignored for dom0 socket services.  Thankfully this feature has not
made it into a release yet.

Fixes: QubesOS/qubes-issues#9066
Fixes: c6ec6ef9ec55 ("Support not passing metadata to socket-based services")