Avoid qrexec-client for VM -> dom0 calls
It isn't needed in this case either.  This saves an execve() and makes
it easier to change the behavior of qrexec-client without breaking
existing qrexec-daemon processes.

For the same reasons as the previous commit, it is critical to avoid
calling exit() in the child process, either directly or indirectly.
Therefore, ensure that none of the code that is now run in the child
process calls exit().  There were places where this was done previously,
so this commit also fixes a bug.