tests: prevent unexpected message combining
If the qrexec daemon is slow enough, it is possible for the standard
output of the service to be sent as one message instead of two.  Prevent
this by waiting for the first message to be received before sending
the service's standard input.  This also requires that the test service
be changed to read its input after producing the output from the
wait-for-session script.