Rip out qubes-rpc-multiplexer
Instead, directly execute the command from C.

All variables with names beginning with QREXEC_ are stripped from the
environment, except for QREXEC_SERVICE_PATH.  This is a change in
behavior compared to the current code.

This is a backwards-incompatible change to
exec_qubes_rpc_if_requested(), which now takes an extra argument.
Therefore, it cannot be backported to R4.2.

Fixes: QubesOS/qubes-issues#9062