Fix flaky qrexec agent tests
Various tests assumed that qrexec would deliver stdout in a single
message.  Qrexec does not make this guarantee: calls to write(2),
send(2), sendmsg(2), etc are not guaranteed to correspond 1-to-1
to MSG_DATA_STDOUT messages on the vchan.  This caused
https://gitlab.com/QubesOS/qubes-core-qrexec/-/jobs/6616564043 to
wrongly fail, even thouggh the code is correct.

Fix this problem by concatenating the payloads of all stdout messages
into a single bytes object before comparing with the expected stdout
value.  Also add a utility function for this, saving a lot of code in
tests.