Check for empty source domain names
The source domain can never be the empty string.  If it _is_ empty, the
most likely cause is a bug in code that is generating the command, such
as two or more spaces after the service descriptor.  This could result
in the parsed source domain being the empty string, while the domain
passed to qubes-rpc-multiplexer is not empty.  Instead of allowing this
tricky-to-debug situation, fail the service call up front.