Ensure that generated commands are well-formed
If the target domain passed on the command line or the normalized target
returned by the policy engine contain a space, ill-formed command lines
would be produced.

To fix this, ensure that both the VM name passed on the command line and
the normalized target returned by the policy engine are non-empty and
contain only printable ASCII characters.  The actual set of valid
characters is smaller, but the input is trusted, so be conservative to
ensure there are no regressions or maintenance problems.  Together with
the previous commit, this change ensures that all commands generated by
qrexec-daemon will be parsed correctly by parse_qubes_rpc_command().

This also ensures that a newline in the VM name passed on the command
line does not result in a malformed request being passed to the policy
daemon.