Reject invalid service targets
If a service target is syntactically incorrect (and would be mangled by
qrexec-daemon), reject it outright instead.  This also adds support for
re-validating data from the policy daemon and the command line, to
ensure that produced commands are correct by construction.  However,
this is turned off in the build, because such changes have a nonzero
risk.

Fixes: QubesOS/qubes-issues#9140