do_fork_exec(): Drop status pipe
The status pipe was closed by fix_fds(), so all of the subsequent
operations on it were operating on a closed file descriptor.  If
exec_func() returned, the subsequent write() would operate on a file
descriptor that might have been reused for another purpose, but none of
the registered execution functions ever returned.  Just call abort() if
they do return.