Add support for exiting on client or service EOF
This adds two new boolean service configuration options:

- exit-on-service-eof: exit when the socket service shuts down its
  output stream for writing.

- exit-on-client-eof: exit when the client shuts down its output stream
  for writing.

The information is passed through an extended qrexec_parsed_command
struct.  New functions are added that avoid the executables having to
access members that are private to libqrexec.

The main use of these features is to emulate the old qubes.ConnectTCP
and qubes.UpdatesProxy services, which already had this behavior due to
the use of socat.

These features are only supported for socket-based services, as
executable services are more complicated and do not have a use case
right now.

Currently, if a service exits due to exit-on-stdin-eof, the empty
MSG_DATA_STDOUT that indicates EOF is not sent.  This is not a problem
because qrexec-client-vm interprets MSG_DATA_EXIT_CODE as also
indicating EOF on stdout and stderr.

Fixes: QubesOS/qubes-issues#9176