Fix build error on redefined _FORTIFY_SOURCE
Leave it for the package builder to define, according to the policy of
specific distribution. And also, do not fight with provided environment,
try to play along instead - drop "override" and custom variables. This
also avoids including $(CFLAGS) twice. This rollback some of the changes
in b883dcb "Makefile improvements"

https://github.com/QubesOS/qubes-builder-archlinux/pull/73#issuecomment-2089227995