Fix SIGUSR1 after stdin_fd closed
Just shut it down for writing, and save it so that it can later be used
for reading.

This also ensures deterministic behavior when --use-stdin-socket is
used.  Previously, the socket would be shut down for reading or writing
but not both, depending on whether reading or writing stopped first.
This could be quite confusing for callers.  Instead, always shut down
the socket for both reading and writing, guaranteeing deterministic
behavior.

This partially reverts a3bbcb5f4314 ("Prefer close() to shutdown()"):
the change to use close() is reverted, but the substantial
simplification of the code is not.

Fixes: a3bbcb5f4314 ("Prefer close() to shutdown()")
Fixes: QubesOS/qubes-issues#9183