Prevent colon in username from causing problems
Previously it could produce invalid qrexec commands.

A username ending with ":nogui" is allowed in the policy engine
response, though not on the command line.  This is because it may be
used by some to work around a limitation of the Windows agent: the
Windows agent handles

    user:nogui:QUBESRPC SERVICE+ARG SOURCE

differently than it handles

    user:QUBESRPC SERVICE+ARG SOURCE

and (unlike the Linux agent) this difference cannot be worked around by
changing service configuration.  R5.0 will block this as well.