daemon: attempt starting target vm only if not running yet
Do not call admin.vm.Start method on every (autostart) call. Do it only
if the target qrexec daemon is not running - or rather, when connecting
to it fails. This makes the happy path of a call a bit quicker (3-5ms in
some basic tests), but more importantly it avoids needlessly checking
with libvirt if the VM is running - which would block if some other
slow operation is running at the time (like, starting another VM).