policy: ignore (with warning) non-allowed default_target=
Instead of denying the call with confusing assert, log a warning
explaining the situation but still display the ask prompt, just without
the default target selected. In some (many?) cases, the intention of
default_target= was to have a prompt just to confirm the operation, not
actual target selection - and in those cases, the intended operation
still won't succeed (as the intended target is not allowed). But in some
cases, user may still want choose an alternative target (for example
after renaming a qube)

QubesOS/qubes-issues#9980