qrexec-agent: remove socket on shutdown
Remove the socket on stop/restart, do not leave stale socket there. This
makes qrexec-client-vm fail more clearly if the agent is not running.
The agent still needs to remove it on startup too, to handle restart
after a crash too, but lets make it a fallback option, not the only
option.

Related to QubesOS/qubes-issues#9982